Refactor : 스케줄 상세 페이지 요청에서 날짜 필드 삭제  및 아이 정보 List에서 단일 정보 변경으로 인한 서비스 로직 수정
