+++ src/main/java/org/guzzing/studayserver/domain/calendar/service/dto/param/AcademyCalendarDetailParam.java
@@ -1,16 +1,8 @@
 package org.guzzing.studayserver.domain.calendar.service.dto.param;
 
-import java.util.List;
-
 public record AcademyCalendarDetailParam(
-        List<ChildrenSchedule> childrenInfos
+        Long childId,
+        Long scheduleId
 ) {
-
-    public record ChildrenSchedule(
-            Long childId,
-            Long scheduleId
-    ) {
-
-    }
 
 }
